rename periodic update -> period summary (code)
