+++ app/v1/controllers/requestParsers/validators/AmendPeriodSummaryValidator.scala
@@ -18,7 +18,7 @@ package v1.controllers.requestParsers.validators
 
 import v1.controllers.requestParsers.validators.validations._
 import v1.models.errors.{MtdError}
-import v1.models.request.amendPeriodic._
+import v1.models.request.amendPeriodSummary._
 
 class AmendPeriodSummaryValidator extends Validator[AmendPeriodSummaryRawData] {
