+++ app/v1/controllers/requestParsers/validators/validations/AmendConsolidatedExpensesValidation.scala
@@ -17,7 +17,7 @@
 package v1.controllers.requestParsers.validators.validations
 
 import v1.models.errors.{MtdError, RuleBothExpensesSuppliedError}
-import v1.models.request.amendPeriodic._
+import v1.models.request.amendPeriodSummary._
 
 object AmendConsolidatedExpensesValidation {
