+++ app/v1/models/request/amendPeriodSummary/AmendPeriodSummaryBody.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v1.models.request.amendPeriodic
+package v1.models.request.amendPeriodSummary
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
