+++ test/v1/mocks/services/MockAmendPeriodSummaryService.scala
@@ -22,7 +22,7 @@ import uk.gov.hmrc.http.HeaderCarrier
 import v1.controllers.EndpointLogContext
 import v1.models.errors.ErrorWrapper
 import v1.models.outcomes.ResponseWrapper
-import v1.models.request.amendPeriodic.AmendPeriodSummaryRequest
+import v1.models.request.amendPeriodSummary.AmendPeriodSummaryRequest
 import v1.services.AmendPeriodSummaryService
 
 import scala.concurrent.{ExecutionContext, Future}
