+++ test/v1/mocks/validators/MockAmendPeriodSummaryValidator.scala
@@ -20,7 +20,7 @@ import org.scalamock.handlers.CallHandler1
 import org.scalamock.scalatest.MockFactory
 import v1.controllers.requestParsers.validators.AmendPeriodSummaryValidator
 import v1.models.errors.MtdError
-import v1.models.request.amendPeriodic.AmendPeriodSummaryRawData
+import v1.models.request.amendPeriodSummary.AmendPeriodSummaryRawData
 
 trait MockAmendPeriodSummaryValidator extends MockFactory {
