+++ test/v1/models/request/amendPeriodSummary/PeriodAllowableExpensesSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v1.models.request.amendPeriodic
+package v1.models.request.amendPeriodSummary
 
 import play.api.libs.json.{JsValue, Json}
 import support.UnitSpec
