+++ test/v1/support/DesResponseMappingSupportSpec.scala
@@ -22,7 +22,7 @@ import utils.Logging
 import v1.controllers.EndpointLogContext
 import v1.models.errors._
 import v1.models.outcomes.ResponseWrapper
-import v1.models.response.createPeriodic.CreatePeriodicResponse
+import v1.models.response.createPeriodSummary.CreatePeriodSummaryResponse
 
 class DesResponseMappingSupportSpec extends UnitSpec {
 
@@ -116,7 +116,7 @@ class DesResponseMappingSupportSpec extends UnitSpec {
   "createPeriodId" should {
     "return a valid periodId" when {
       "given two valid dates" in {
-        val responseModel = CreatePeriodicResponse("2017-09-09_2017-09-09")
+        val responseModel = CreatePeriodSummaryResponse("2017-09-09_2017-09-09")
         mapping.createPeriodId(ResponseWrapper(correlationId, ()), "2017-09-09", "2017-09-09") shouldBe
           Right(ResponseWrapper(correlationId, responseModel))
       }
